API: move several functions to write.py, changing API
